requirements.txt
Flask==3.0.3
psycopg==3.2.3
python-dotenv==1.0.1
nltk==3.8.1
